Backend/gitlab/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.3
billiard==4.1.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
DateTime==5.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-cron==0.6.0
django-crontab==0.7.1
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.0.1
django-templated-mail==1.1.1
django4-background-tasks==1.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
et-xmlfile==1.1.0
fire==0.5.0
firebase-admin==6.2.0
future==0.18.2
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
msgpack==1.0.6
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
prettytable==3.7.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.3
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
pyzk==0.9
pyzkaccess==1.0
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
sentry-sdk==1.28.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.2.1
termcolor==2.3.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
